Moved PartSynchronizer into timed/ subpackage and fixed it.
Now synchronizer emits 2-tick pulse as all other gates.
