+++ src/main/java/moe/nightfall/vic/integratedcircuits/cp/CircuitPart.java
@@ -11,7 +11,6 @@ import moe.nightfall.vic.integratedcircuits.Constants;
 import moe.nightfall.vic.integratedcircuits.cp.part.PartIOBit;
 import moe.nightfall.vic.integratedcircuits.cp.part.PartMultiplexer;
 import moe.nightfall.vic.integratedcircuits.cp.part.PartNull;
-import moe.nightfall.vic.integratedcircuits.cp.part.PartSynchronizer;
 import moe.nightfall.vic.integratedcircuits.cp.part.PartTorch;
 import moe.nightfall.vic.integratedcircuits.cp.part.PartTunnel;
 import moe.nightfall.vic.integratedcircuits.cp.part.PartWire;
@@ -35,6 +34,7 @@ import moe.nightfall.vic.integratedcircuits.cp.part.timed.PartRandomizer;
 import moe.nightfall.vic.integratedcircuits.cp.part.timed.PartRepeater;
 import moe.nightfall.vic.integratedcircuits.cp.part.timed.PartSequencer;
 import moe.nightfall.vic.integratedcircuits.cp.part.timed.PartStateCell;
+import moe.nightfall.vic.integratedcircuits.cp.part.timed.PartSynchronizer;
 import moe.nightfall.vic.integratedcircuits.cp.part.timed.PartTimer;
 import moe.nightfall.vic.integratedcircuits.misc.CraftingAmount;
 import moe.nightfall.vic.integratedcircuits.misc.PropertyStitcher;
